move money life and teamsize info
