+++ app/public/src/pages/component/icons/life.tsx
@@ -5,7 +5,7 @@ export function Life(props: { value?: number | string }) {
   return (
     <>
       {props.value && (
-        <span style={{ verticalAlign: "middle" }}>{props.value}</span>
+        <span style={{ verticalAlign: "middle", flex: "1" }}>{props.value}</span>
       )}
       <img className="icon-life" src="/assets/ui/heart.png" alt="❤" />
     </>
