+++ app/public/src/pages/component/icons/money.tsx
@@ -5,7 +5,7 @@ export function Money(props: { value?: number | string }) {
   return (
     <>
       {props.value && (
-        <span style={{ verticalAlign: "middle" }}>{props.value}</span>
+        <span style={{ verticalAlign: "middle", flex: "1" }}>{props.value}</span>
       )}
       <img className="icon-money" src="/assets/ui/money.svg" alt="$" />
     </>
